refactor: move main application logic into separate `lib.rs` file
